chore(flow-edit): split test modules to honour 200 LOC cap
Codex P3 round-2 finding: ``flow_edit_test.py`` was 346 LOC and
``flow_edit_helpers_test.py`` was 210 LOC, both over AGENTS.md's hard
cap.  Split by responsibility:

* ``flow_edit_helpers_test.py`` (134) — schedule, CFG dispatch, packing,
  noise draw.
* ``flow_edit_helpers_velocity_test.py`` (90) — clamp, EMA, momentum
  snapshot/restore.
* ``flow_edit_test.py`` (191) — sampling-loop contract: window routing,
  determinism, error guards, APG-momentum regression, EMA regression.
* ``flow_edit_pipeline_test.py`` (87) — pipeline-level conditioning-
  hint forwarding regression.
* ``_flow_edit_test_support.py`` (75) — shared ``StubDecoder`` /
  ``StubModel`` / ``make_inputs`` (underscored to skip discovery).

All 27 tests still pass.  No source changes — purely a test-layout
refactor to satisfy the module-size policy.
