+++ acestep/models/common/flow_edit_helpers_test.py
@@ -1,4 +1,9 @@
-"""Unit tests for ``flow_edit_helpers`` primitives (#1156)."""
+"""Schedule, CFG, packing, and noise unit tests for ``flow_edit_helpers``.
+
+Velocity / momentum helpers live in
+``flow_edit_helpers_velocity_test.py`` (split per the AGENTS.md
+200 LOC module cap).
+"""
 
 import unittest
 
@@ -7,13 +12,9 @@ import torch
 from acestep.models.common.apg_guidance import MomentumBuffer
 from acestep.models.common.flow_edit_helpers import (
     apply_cfg_branch,
-    apply_velocity_clamp,
-    apply_velocity_ema,
     build_timestep_schedule,
     draw_fwd_noise,
     pack_for_cfg,
-    restore_and_advance_momentum,
-    snapshot_momentum,
 )
 
 
@@ -25,7 +26,6 @@ class BuildTimestepScheduleTests(unittest.TestCase):
         self.assertEqual(t.shape, (9,))
         self.assertAlmostEqual(t[0].item(), 1.0)
         self.assertAlmostEqual(t[-1].item(), 0.0)
-        # Even spacing for shift=1
         diffs = (t[:-1] - t[1:]).tolist()
         self.assertTrue(all(abs(d - diffs[0]) < 1e-6 for d in diffs))
 
@@ -59,94 +59,21 @@ class ApplyCfgBranchTests(unittest.TestCase):
 
     def test_no_cfg_passes_through_unchanged(self):
         out = apply_cfg_branch(self.pred_cond, do_cfg=False, apply_cfg_now=True,
-                                guidance_scale=7.0, momentum_buffer=MomentumBuffer())
+                               guidance_scale=7.0, momentum_buffer=MomentumBuffer())
         self.assertTrue(torch.equal(out, self.pred_cond))
 
     def test_cfg_outside_interval_returns_cond_chunk(self):
         out = apply_cfg_branch(self.pred_packed, do_cfg=True, apply_cfg_now=False,
-                                guidance_scale=7.0, momentum_buffer=MomentumBuffer())
+                               guidance_scale=7.0, momentum_buffer=MomentumBuffer())
         self.assertTrue(torch.equal(out, self.pred_cond))
 
     def test_cfg_inside_interval_uses_apg(self):
         buf = MomentumBuffer()
         out = apply_cfg_branch(self.pred_packed, do_cfg=True, apply_cfg_now=True,
-                                guidance_scale=7.0, momentum_buffer=buf)
+                               guidance_scale=7.0, momentum_buffer=buf)
         # APG amplifies the orthogonal component, so the result must differ
         # from a no-op pass-through of the cond chunk.
         self.assertFalse(torch.equal(out, self.pred_cond))
-
-
-class ApplyVelocityClampTests(unittest.TestCase):
-    """Velocity-norm clamp must be a no-op when threshold is 0."""
-
-    def test_threshold_zero_is_noop(self):
-        torch.manual_seed(0)
-        vt = torch.randn(2, 4, 8)
-        xt = torch.randn(2, 4, 8)
-        self.assertTrue(torch.equal(apply_velocity_clamp(vt, xt, 0.0), vt))
-
-    def test_clamp_reduces_outlier_norms(self):
-        vt = torch.full((1, 4, 8), 100.0)
-        xt = torch.full((1, 4, 8), 1.0)
-        out = apply_velocity_clamp(vt, xt, 2.0)
-        out_norm = torch.norm(out, dim=(1, 2)).item()
-        xt_norm = torch.norm(xt, dim=(1, 2)).item()
-        self.assertAlmostEqual(out_norm, 2.0 * xt_norm, places=2)
-
-
-class ApplyVelocityEmaTests(unittest.TestCase):
-    """EMA smoothing: per-timestep blend, not per-MC-draw (regression for codex P2)."""
-
-    def test_ema_no_prev_returns_vt(self):
-        vt = torch.randn(1, 4, 8)
-        self.assertTrue(torch.equal(apply_velocity_ema(vt, None, 0.5), vt))
-
-    def test_ema_zero_factor_returns_vt(self):
-        vt = torch.randn(1, 4, 8)
-        prev = torch.randn(1, 4, 8)
-        self.assertTrue(torch.equal(apply_velocity_ema(vt, prev, 0.0), vt))
-
-    def test_ema_with_prev_blends(self):
-        vt = torch.ones(1, 4, 8)
-        prev = torch.zeros(1, 4, 8)
-        out = apply_velocity_ema(vt, prev, 0.3)
-        # 0.7*1 + 0.3*0 = 0.7
-        self.assertTrue(torch.allclose(out, torch.full_like(vt, 0.7)))
-
-
-class MomentumSnapshotTests(unittest.TestCase):
-    """Snapshot/restore must isolate APG momentum across MC draws."""
-
-    def test_snapshot_then_advance_matches_single_update(self):
-        """One pre-step snapshot + one final update == one direct update."""
-        buf_a, buf_b = MomentumBuffer(), MomentumBuffer()
-        buf_a.update(torch.tensor([1.0, 2.0]))
-        # Snapshot AFTER the prev-step state.
-        snap_a, snap_b = snapshot_momentum(buf_a, buf_b)
-        # Simulate inner loop "polluting" buf_a with a diff that should NOT
-        # carry forward — we restore-and-advance with avg_diff outside.
-        buf_a.update(torch.tensor([99.0, 99.0]))
-        # Reference: apply only the avg_diff once from the snapshot.
-        ref = MomentumBuffer()
-        ref.update(torch.tensor([1.0, 2.0]))
-        ref.update(torch.tensor([3.0, 4.0]))
-        # Restore and advance once with the avg.
-        restore_and_advance_momentum(
-            buf_a, buf_b, snap_a, snap_b,
-            avg_diff_src=torch.tensor([3.0, 4.0]),
-            avg_diff_tar=torch.tensor([5.0, 6.0]),
-        )
-        self.assertTrue(torch.equal(buf_a.running_average, ref.running_average))
-
-    def test_no_cfg_diff_just_rolls_back(self):
-        """When CFG is inactive (avg_diff is None) the buffer rolls back."""
-        buf_a, buf_b = MomentumBuffer(), MomentumBuffer()
-        buf_a.update(torch.tensor([1.0]))
-        snap_a, snap_b = snapshot_momentum(buf_a, buf_b)
-        # Simulate pollution.
-        buf_a.update(torch.tensor([42.0]))
-        restore_and_advance_momentum(buf_a, buf_b, snap_a, snap_b, None, None)
-        self.assertTrue(torch.equal(buf_a.running_average, snap_a))
 
 
 class DrawFwdNoiseTests(unittest.TestCase):
@@ -158,7 +85,7 @@ class DrawFwdNoiseTests(unittest.TestCase):
         a = draw_fwd_noise(self.SHAPE, None, torch.device("cpu"), torch.float32)
         b = draw_fwd_noise(self.SHAPE, None, torch.device("cpu"), torch.float32)
         self.assertEqual(a.shape, self.SHAPE)
-        self.assertFalse(torch.equal(a, b))  # different draws
+        self.assertFalse(torch.equal(a, b))
 
     def test_single_generator_reproducible(self):
         g1 = torch.Generator().manual_seed(42)
@@ -171,7 +98,6 @@ class DrawFwdNoiseTests(unittest.TestCase):
         gens = [torch.Generator().manual_seed(i) for i in (1, 2, 3)]
         out = draw_fwd_noise(self.SHAPE, gens, torch.device("cpu"), torch.float32)
         self.assertEqual(out.shape, self.SHAPE)
-        # Each sample should differ (different seeds).
         self.assertFalse(torch.equal(out[0], out[1]))
         self.assertFalse(torch.equal(out[1], out[2]))
 
@@ -196,12 +122,10 @@ class PackForCfgTests(unittest.TestCase):
 
     def test_cfg_doubles_batch_and_uses_null(self):
         out = pack_for_cfg(self.enc_hs, self.enc_am, self.ctx, self.attn, self.null, do_cfg=True)
-        # Batch dim doubled
         self.assertEqual(out[0].shape[0], 4)
         self.assertEqual(out[1].shape[0], 4)
         self.assertEqual(out[2].shape[0], 4)
         self.assertEqual(out[3].shape[0], 4)
-        # First half = original, second half = null-broadcast
         self.assertTrue(torch.equal(out[0][:2], self.enc_hs))
         self.assertTrue(torch.equal(out[0][2:], self.null.expand_as(self.enc_hs)))
